-Added BookMark functionality -Updated Profile Screen and Edit Profile -Add,Remove,Rate Book

+++ lib/util/size_config.dart
@@ -19,7 +19,8 @@ class SizeConfig {
 double getProportionateScreenHeight(double inputHeight) {
   double screenHeight = SizeConfig.screenHeight;
   // 812 is the layout height that designer use
-  return (inputHeight / 812.0) * screenHeight;
+  //convert result to double
+  return ((inputHeight / 812.0) * screenHeight).toDouble();
 }
 
 // Get the proportionate height as per screen size
